Fixed Bug on Similar Item Recommender Notebook
Earlier, when working with smaller sample size, we use pandas.sample(X)
which actually create X number of random samples. However, when we
try to map them back to images on the folder, we assumed they're
in 1-X ordering, which wasn't the case. So the solution is not to use
.sample(X), but [0:X) instead